moves tests too, and move utility function
